requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.28.1
python-dotenv==1.0.0
plotly==5.17.0
streamlit-chat==0.1.1
torch==2.0.1
transformers==4.35.0
nltk==3.8.1